% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{localize}
\alias{localize}
\alias{localizeSingle}
\alias{localizeMultiple}
\title{Localize detected sounds}
\usage{
localize(
  wavList,
  coordinates,
  margin = 10,
  zMin = -1,
  zMax = 20,
  resolution = 1,
  F_Low = 2000,
  F_High = 8000,
  tempC = 15,
  soundSpeed = NULL,
  plot = TRUE,
  locFolder = NULL,
  jpegName = "000.jpeg",
  InitData = NULL,
  keep.InitData = TRUE,
  keep.SearchMap = FALSE
)

localizeSingle(
  st,
  index,
  tempC = 15,
  plot = TRUE,
  InitData = NULL,
  keep.InitData = TRUE,
  keep.SearchMap = FALSE
)

localizeMultiple(st, indices = "all", tempC = 15, plot = TRUE, InitData = NULL)
}
\arguments{
\item{wavList}{list of Wave objects. The name of the Wave objects MUST be
present in the coordinates data.frame.}

\item{coordinates}{data.frame. Must contain four required columns:
column Station contains a character string with names of each recording
station, while Easting, Northing and Elevation contain the x, y, and z
coordinates of the station, in meters (E.g. UTM coordinates).}

\item{margin, zMin, zMax, resolution}{Arguments describing the area to be searched
for sound sources. Passed to \code{\link{makeSearchMap}}.}

\item{F_Low, F_High}{Numeric. The low and high frequency, in Hz, of the sound
to be localized.}

\item{tempC}{Numeric. Temperature in degrees C, which is used to calculate
the speed of sound in air using the equation 331.45*sqrt(1+tempC/273.15).}

\item{soundSpeed}{Numeric. The speed of sound in meters per second. Default is
NULL, in which case the speed of sound is calculated based on the specified
temperature (assuming the transmission medium is air). If soundSpeed is
specified, the tempC value is over-ridden.}

\item{plot}{Logical. Whether to plot jpegs.}

\item{locFolder}{Character. File path to the folder where localization jpegs
(heatmaps and spectrograms) are to be created. Only required if plot = TRUE.}

\item{jpegName}{Character. Name of the jpeg, ending in extension .jpeg.
Only required if plot = TRUE.}

\item{InitData}{List. An InitData list created by running localization with
keep.InitData = TRUE. Providing an InitData list saves computation time,
but is only possible if the SearchGrid and stations used for localization
remain unchanged. Default is NULL, which means the InitData will be
calculated anew.}

\item{keep.InitData}{Logical. Whether to store the InitData list.}

\item{keep.SearchMap}{Logical. Whether to keep the SearchMap list with
power estimates and coordinates of each grid cell.
Should only be set to TRUE if the SearchMap is needed
for some other reason (e.g. making a publication-ready figure or
conducting more involved analysis with overlapping sources, etc.).}

\item{st}{List. Localization settings object generated using
\code{\link{processSettings}}. Only needed for \code{localizeSingle} or
\code{localizeMultiple}.}

\item{index, indices}{Numeric or 'all'. Indices to be localized within a detection file.
Setting to 1 localizes the first row, c(7:10) localizes rows 7-10, and 'all'
localizes all rows (ignoring rows that have no entry in the Station1 column).}
}
\value{
List, containing the location of the sound source (global maximum),
    and optionally the InitData and SearchMap lists.
}
\description{
\code{localize} and the two related functions \code{localizeSingle} and \code{localizeMultiple}
are the basic functions for localizing sounds. They take audio data as inputs, alongside relevant
 metadata (e.g. coordinates and a variety of settings), and estimate the location of the
 dominant sound source. The \code{localize} function takes as arguments the minimal
amount of information needed for localization. Localization is conducted on the
full duration of the Wave objects in wavList. Effectively this means the user
must wrangle the data and clip the audio themselves, but this affords the greatest
flexibility in terms of how the user chooses to organize their data.
The \code{localizeSingle} and \code{localizeMultiple} functions, in contrast,
automate much of the data wrangling process, but require data to be organized in a
very specific way (e.g. folder structure, file structures). Thus, the latter two
functions tradeoff flexibility for increased automation. All three functions use the
same underlying localization algorithm (\code{localizeSingle} and \code{localizeMultiple}
pass their data to \code{localize} after the data has been wrangled).
}
