% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSRP_Init.R
\name{MSRP_Init}
\alias{MSRP_Init}
\title{Create InitData.}
\usage{
MSRP_Init(NodeInfo, SearchMap, Para, LevelFlag)
}
\arguments{
\item{NodeInfo}{List with elements Num, Pos.}

\item{SearchMap}{List with elements XDen, YDen, ZDen, XMap, YMap, ZMap}

\item{Para}{List with Fs, Vc (speed of sound), and DataLen}

\item{LevelFlag}{Integer. Only value currently supported is 2.}
}
\value{
List.
}
\description{
Internal function which creates the InitData list.
}
\author{
Tim Huang.
}
