% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rij_GCC.R
\name{Rij_GCC}
\alias{Rij_GCC}
\title{Generalized cross-correlation.}
\usage{
Rij_GCC(data1, data2, Para)
}
\arguments{
\item{data1, data2}{Wave samples.}

\item{Para}{List with GCCMethod, FL, FH, Fs}
}
\value{
Numeric vector.
}
\description{
Internal function that calculates the generalized cross correlation.
}
