% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationSpec.R
\name{validationSpec}
\alias{validationSpec}
\title{Create validation spectrograms.}
\usage{
validationSpec(
  wavList,
  coordinates,
  locationEstimate,
  from = NULL,
  to = NULL,
  tempC = 15,
  soundSpeed = NULL,
  F_Low,
  F_High
)
}
\arguments{
\item{wavList}{list of Wave objects. The name of the Wave objects MUST be
present in the coordinates data.frame.}

\item{coordinates}{data.frame. Must contain four required columns:
column Station contains a character string with names of each recording
station, while Easting, Northing and Elevation contain the x, y, and z
coordinates of the station, in meters (E.g. UTM coordinates).}

\item{locationEstimate}{Dataframe with one row containing columns Easting, Northing and Elevation,
specifying the estimated location of the sound source.}

\item{from, to}{Numeric. The portion of the wavs to plot.}

\item{tempC}{Numeric. The ambient temperature in celsius, which is used to calculate the speed
of sound in air if none is specified.}

\item{soundSpeed}{Numeric. The speed of sound. If not provided, tempC will be used to calculate
the speed of sound in air.}

\item{F_Low, F_High}{Numeric. The low and high frequency, in Hz, of the sound
to be localized.}
}
\description{
This function is used inside the \code{localize} function to create the panels of
synchronized spectrograms for manual review.
}
