% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validSpectro.R
\name{validSpectro}
\alias{validSpectro}
\title{Create validation spectrograms.}
\usage{
validSpectro(ListOfData)
}
\arguments{
\item{ListOfData}{List containing several required elements to make a single panel of the plot.}
}
\description{
This function is used inside the \code{localize} function to create the panels of
synchronized spectrograms for manual review.
}
\details{
The plot is created using the `imagep` function from the `oce` package, in the active
    graphical device.
}
