% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locaR-package.R
\docType{package}
\name{locaR-package}
\alias{locaR}
\alias{locaR-package}
\title{locaR: A Set of Tools for Sound Localization}
\description{
A set of functions and tools to conduct acoustic source localization, as well as organize and check localization data and results. The localization functions implement the modified steered response power algorithm described by Cobos et al. (2010) \doi{10.1109/LSP.2010.2091502}.
}
\details{
The locaR package contains functions for localizing sounds using R.
Localizations are carried out using the modified steered response power
algorithm of Cobos et al. (2011) which carries out a grid-search to find
the location in three dimensions where a sound was most likely to have
originated.
}
\references{
Cobos, M., Martí, A., & J.J. López. 2011. A modified SRP-PHAT functional for robust real-time sound source localization with scalable spatial sampling. IEEE Signal Processing Letters. 18:71-74. doi:10.1109/LSP.2010.2091502.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rhedley/locaR}
  \item Report bugs at \url{https://github.com/rhedley/locaR/issues}
}

}
\author{
\strong{Maintainer}: Richard Hedley \email{rwhedley@gmail.com}

Authors:
\itemize{
  \item Marcus Becker \email{mabecker@ualberta.ca}
  \item Tim Huang
}

}
\keyword{internal}
