% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locaR.R
\docType{package}
\name{locaR}
\alias{locaR}
\title{locaR: Sound localization in R.}
\description{
The locaR package contains functions for localizing sounds using R.
Localizations are carried out using the modified steered response power
algorithm of Cobos et al. (2011) which carries out a grid-search to find
the location in three dimensions where a sound was most likely to have
originated.
}
\section{locaR functions}{

The locaR functions...
}

