% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSRP_RIJ_HT.R
\name{MSRP_RIJ_HT}
\alias{MSRP_RIJ_HT}
\alias{MSRP_HT_Level2}
\title{Internal function for localization.}
\usage{
MSRP_RIJ_HT(NodeInfo, SearchMap, Data, Para, LevelFlag, InitData)

MSRP_HT_Level2(NodeInfo, SearchMap, Data, Para, InitData)
}
\arguments{
\item{NodeInfo}{List with elements Num, Pos.}

\item{SearchMap}{List with elements XDen, YDen, ZDen, XMap, YMap, ZMap}

\item{Data}{Matrix containing the wave samples.}

\item{Para}{List with Fs, Vc (speed of sound), and DataLen}

\item{LevelFlag}{Integer. Only value currently supported is 2.}

\item{InitData}{List. Created with the MSRP_Init function.}
}
\value{
List.
}
\description{
This function uses the InitData and other info to calculate the likelihood of sound
sources coming from each location. Note: the LevelFlag argument is currently redundant because
there is only one option. Similarly, the MSRP_HT_Level2 function could be rolled into
the MSRP_RIJ_HT function in the future, but for now is kept separate.
}
\author{
Tim Huang.
}
