% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseWAFileNames.R
\name{parseWAFileNames}
\alias{parseWAFileNames}
\title{Parse Wildlife Acoustics-type file names.}
\usage{
parseWAFileNames(filenames)
}
\arguments{
\item{filenames}{Character vector of file names.}
}
\value{
A data frame with prefix, channels, date, time and extension information.
}
\description{
This function parses the information in file names that are structured according to Wildlife Acoustics'
naming convention. Specifically, the format prefix_date_time.wav or prefix_mic_date_time.wav.
}
\examples{
parseWAFileNames("Ex-4_0+1_20200617$090000.wav") #One file.
parseWAFileNames(c("Ex-4_0+1_20200617$090000.wav", "Ex-4_0+1_20200618$053000.wav")) #Multiple files.
}
